backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==3.2.3
django-cors-headers==3.13.0
django-filter==23.5
django-sql-utils==0.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
exceptiongroup==1.2.1
executing==2.0.1
filetype==1.2.0
gunicorn==20.1.0
idna==3.7
ipython==8.18.1
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.11.1
